Corrected Validation Feature and added create Products

+++ src/main/java/com/vasanth/Ecommerce_Backend/model/Product.java
@@ -34,7 +34,7 @@ public class Product {
     private Integer quantity;
     private double price;
     private double discount;
-    private Integer specialPrice;
+    private double specialPrice;
 
     @ManyToOne
     @JoinColumn(name = "category_id")
